+++ index.js
@@ -3,7 +3,7 @@ require('babel-polyfill');
 
 const express = require('express');
 const app = express();
-const routes = require('./routes.js');
+const routes = require('./src/routes.js');
 const connection = require('./connection');
 
 module.exports = app;
