+++ src/routes/theses.js
@@ -1,0 +1,17 @@
+const router = require('express').Router();
+const bodyParser = require('body-parser');
+const output = require('../output');
+const thesisController = require('../controllers/ThesisController');
+const jsonParser = bodyParser.json()
+const urlencodedParser = bodyParser.urlencoded({ extended: false })
+
+
+router.get('/', (req, res) => {
+    thesisController.getAllTheses(req, res);
+});
+
+router.get('/:id', (req, res) => {
+    thesisController.getThesisById(req, res);
+});
+
+module.exports = router;
